# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Goat Farm Management API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/simulation/defaults/breeds": {
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "List Breeds",
        "description": "Available breed presets and production systems.",
        "operationId": "list_breeds_api_simulation_defaults_breeds_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreedsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/defaults": {
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "Breed Defaults",
        "description": "Default assumptions for a breed + production system (400 on unknown).",
        "operationId": "breed_defaults_api_simulation_defaults_get",
        "parameters": [
          {
            "name": "breed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "osmanabadi",
              "title": "Breed"
            }
          },
          {
            "name": "system",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "stall_fed",
              "title": "System"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationAssumptions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/herd-snapshot": {
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "Herd Snapshot",
        "description": "Group the farm's ACTIVE animals into simulation starting cohorts:\nkid 0-2 m, weaner 3-5 m, grower 6 m up to breeding age, adult at breeding\nage (doe threshold = the breed's age-at-first-breeding, buck at 12 m;\nunknown age counts as adult).",
        "operationId": "herd_snapshot_api_simulation_herd_snapshot_get",
        "parameters": [
          {
            "name": "breed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "osmanabadi",
              "title": "Breed"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HerdSnapshotOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/run": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Run Adhoc",
        "description": "Run a simulation from posted assumptions (no persistence).",
        "operationId": "run_adhoc_api_simulation_run_post",
        "parameters": [
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/scenarios": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Create Scenario",
        "operationId": "create_scenario_api_simulation_scenarios_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioCreateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "List Scenarios",
        "operationId": "list_scenarios_api_simulation_scenarios_get",
        "parameters": [
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScenarioOut"
                  },
                  "title": "Response List Scenarios Api Simulation Scenarios Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/scenarios/compare": {
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "Compare Scenarios",
        "description": "Run 2+ stored scenarios deterministically side by side (``ids=1,2``).\n\nDuplicate ids are collapsed (a repeated id must not re-run a simulation);\nmore than MAX_COMPARE_IDS distinct ids is a 400.",
        "operationId": "compare_scenarios_api_simulation_scenarios_compare_get",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ids"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioCompareOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/scenarios/{scenario_id}": {
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "Get Scenario",
        "operationId": "get_scenario_api_simulation_scenarios__scenario_id__get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scenario Id"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "simulation"
        ],
        "summary": "Update Scenario",
        "operationId": "update_scenario_api_simulation_scenarios__scenario_id__patch",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scenario Id"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "simulation"
        ],
        "summary": "Delete Scenario",
        "operationId": "delete_scenario_api_simulation_scenarios__scenario_id__delete",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scenario Id"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulation/scenarios/{scenario_id}/run": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Run Scenario",
        "description": "Run a stored scenario's assumptions (optionally with MC / sensitivity).",
        "operationId": "run_scenario_api_simulation_scenarios__scenario_id__run_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scenario Id"
            }
          },
          {
            "name": "monte_carlo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Monte Carlo"
            }
          },
          {
            "name": "sensitivity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Sensitivity"
            }
          },
          {
            "name": "x-farm-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Farm-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BreedsOut": {
        "properties": {
          "breeds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Breeds"
          },
          "systems": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Systems"
          }
        },
        "type": "object",
        "required": [
          "breeds",
          "systems"
        ],
        "title": "BreedsOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HerdSnapshotOut": {
        "properties": {
          "does": {
            "type": "integer",
            "title": "Does"
          },
          "bucks": {
            "type": "integer",
            "title": "Bucks"
          },
          "f_kids": {
            "type": "integer",
            "title": "F Kids"
          },
          "f_weaners": {
            "type": "integer",
            "title": "F Weaners"
          },
          "f_growers": {
            "type": "integer",
            "title": "F Growers"
          },
          "m_kids": {
            "type": "integer",
            "title": "M Kids"
          },
          "m_weaners": {
            "type": "integer",
            "title": "M Weaners"
          },
          "m_growers": {
            "type": "integer",
            "title": "M Growers"
          },
          "total_head": {
            "type": "integer",
            "title": "Total Head"
          }
        },
        "type": "object",
        "required": [
          "does",
          "bucks",
          "f_kids",
          "f_weaners",
          "f_growers",
          "m_kids",
          "m_weaners",
          "m_growers",
          "total_head"
        ],
        "title": "HerdSnapshotOut",
        "description": "The farm's ACTIVE animals grouped into simulation starting cohorts."
      },
...
    }
  }
}